Vim disable plugin auto-deletion, better popups, better taskwarrior alternative line
I'll explain only about the plugin auto-deletion: I like to comment
plugins I was experimenting but I don't want to use anymore on `./sync`,
but my plugins list was starting to get a bit too messy; because of
that, I made a function that checks if there are plugin folders that
weren't added via the commands and deletes them. The only exception are
symlinks (usually pointing to local git repos in my machine).