tty coloring and flameshot script
* Added gruvbox theme for the linux terminal (when out of any X server).
I'm still trying this out to see if it's worth it. Code is on zshrc.

* Added support for flameshot in my xprofile & sxhkdrc; I've also made a
`flamedash´ script that helps taking screenshots on automated places.
And I was trying to use scrot before but gave up (that's why there is a
new file on local/bin/trash called scshot - it was an attempt to make a
script with scrot).