Changed downloads folder, replaced a vim completion plugin
* Added an export: XDG_DOWNLOAD_DIR=~/inbox. I don't know if it'll work
for what I want (change default download folder for gtk popups), so I'm
still testing it.

* I use two completion plugins for vim: deoplete.nvim (when I have
python, I like it because of the fuzzy finding), and I used to have
VimCompletesMe for other situations, but I found a better one called
vim-auto-popmenu and I'm now using it for the second case.