Finally made a good opener script
Made a new opener script. It is based off nuke, which is a script
available on the nnn GitHub repository. I simplified it and it looks way
simpler now, in my opinion. I also made another script, openfork, to
complement the opener. It basically opens the file specified in $1
asynchronously and dettaches, making it possible for me to, for example,
open an image in sxiv while in nnn and closing the terminal window with
nnn, while letting the sxiv window still open.

I've been wanting to do this for quite some time, and I'm happy I
finally got to finish it.